requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
qdrant-client>=1.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-qdrant>=0.1.0
openai>=1.3.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
PyPDF2>=3.0.0
python-docx>=0.8.11
ragas>=0.1.0
datasets>=2.14.0
sentence-transformers>=2.2.0
spacy>=3.6.0
pandas>=1.5.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.15.0
requests>=2.31.0
